requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn 
pydantic>=2.0.0,<3.0.0    
pydantic-settings>=2.0.0,<3.0.0
python-multipart

# LangChain v0.3 dependencies
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# OpenAI
openai

# Google API dependencies for Gmail
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Vector database and embeddings
faiss-cpu==1.7.4
numpy==1.24.3

# Redis for caching
aioredis==2.0.1

# Environment and configuration
python-dotenv==1.0.0

# Async support

aiofiles==23.2.1

# Text processing
tiktoken

# Logging and monitoring
structlog==23.2.0

# HTTP client
httpx==0.25.2

# Email parsing
email-validator==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0